# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/audit-log:
    get:
      operationId: getAuditLog
      tags: [dashboard]
      summary: Get recent audit log entries
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: action
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"

  # ─── PII ────────────────────────────────────────────────────────────────────
...
  /admin/db-connections:
    get:
      operationId: listDbConnections
      tags: [workflow]
      summary: List all DB connections (admin)
      responses:
        "200":
          description: List of DB connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DbConnection"
    post:
      operationId: createDbConnection
      tags: [workflow]
      summary: Create a DB connection (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbConnectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbConnection"
  /admin/db-connections/{id}:
    put:
      operationId: updateDbConnection
      tags: [workflow]
      summary: Update a DB connection (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbConnectionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbConnection"
    delete:
      operationId: deleteDbConnection
      tags: [workflow]
      summary: Delete a DB connection (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/db-connections/{id}/test:
    post:
      operationId: testDbConnection
      tags: [workflow]
      summary: Test a DB connection (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbConnectionTestResult"

  # ─── API DOCS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AuditLogEntry:
      type: object
      required: [id, userId, userEmail, action, details, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        userEmail:
          type: ["string", "null"]
        action:
          type: string
        details:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        resourceType:
          type: ["string", "null"]
        resourceId:
          type: ["string", "null"]
        fieldName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      required: [entries, total, page, pageSize]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ─── PII ─────────────────────────────────────────────────────────────────
...
    DbConnection:
      type: object
      required: [id, name, type, host, port, dbName, schemaName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [backoffice, trading]
        host:
          type: string
        port:
          type: integer
        dbName:
          type: string
        schemaName:
          type: string
        fetchQuery:
          type: ["string", "null"]
        outputFilePath:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        lastTestedAt:
          type: ["string", "null"]
          format: date-time
        lastTestSuccess:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DbConnectionInput:
      type: object
      required: [name, type, host, dbName, username, password]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [backoffice, trading]
        host:
          type: string
        port:
          type: integer
          default: 5432
        dbName:
          type: string
        schemaName:
          type: string
          default: public
        username:
          type: string
        password:
          type: string
        outputFilePath:
          type: string
        fetchQuery:
          type: string
    DbConnectionTestResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: ["string", "null"]
...